Remove old Rope impl, migrate tests to src/rope.rs
There's more work to do here for sure, but this is a small step in the
right direction.
