+++ rust/rope/examples/ropetoy.rs
@@ -26,7 +26,7 @@ fn main() {
         a.edit_str(l, l, &(i.to_string() + "\n"));
     }
     let l = a.len();
-    for s in a.clone().slice(1000, 3000).iter_chunks() {
+    for s in a.clone().iter_chunks(1000, 3000) {
         println!("chunk {:?}", s);
     }
     a.edit_str(1000, l, "");
